+++ Project_wsl_2026/Software/main_tennsy4.0/lib/Kicker/kicker.cpp
@@ -2,21 +2,21 @@
 
 /* キッカーで蹴ったり貯めたりする */
 
-int16_t fet1_kick_pin = 0;   // fet1につないだピン
-int16_t fet2_charge_pin = 0; // fet2につないだピン
+int fet1_kick_pin = 0;   // fet1につないだピン
+int fet2_charge_pin = 0; // fet2につないだピン
 
-int16_t kicker_cooldown_time = 1000; // クールダウンタイム
+int kicker_cooldown_time = 1000; // クールダウンタイム
 
 Timer my_cooldown_timer;       // クールダウン用タイマー
 bool kicker_first_call = true; // 最初の呼び出しかどうか
 
-void kicker_set_fetpin(int16_t kick_pin, int16_t charge_pin)
+void kicker_set_fetpin(int kick_pin, int charge_pin)
 {
     fet1_kick_pin = kick_pin;
     fet2_charge_pin = charge_pin;
 }
 
-void kicker_init(int16_t cooldown_time)
+void kicker_init(int cooldown_time)
 {
     pinMode(fet1_kick_pin, OUTPUT);
     pinMode(fet2_charge_pin, OUTPUT);
@@ -54,9 +54,9 @@ void kicker_kick(bool kick_signal)
 
 /* キャッチセンサー */
 
-static int16_t catchsensor_pin = 0;
+static int catchsensor_pin = 0;
 
-void catchsensor_init(int16_t pin)
+void catchsensor_init(int pin)
 {
     catchsensor_pin = pin;
     pinMode(catchsensor_pin, INPUT);
