+++ Project_wsl_2026/Software/main_tennsy4.0/lib/Motors/compute/motors-const_ver.cpp
@@ -3,14 +3,14 @@
 /*すべてのモータの合力が常に指定した値になるようにモーターを制御する*/
 
 /*モーターはすべて正で、まっすぐ進むようにする*/
-int8_t motor_pid_sign[4] = {1, -1, -1, 1}; // モータを回転させるための符号を格納
+int motor_pid_sign[4] = {1, -1, -1, 1}; // モータを回転させるための符号を格納
 
-int16_t motor_deg[4];             // モータの設置角度を格納用
-int16_t motor_move_power_main[4]; // モータのスケーリングされた出力格納用(移動のための出力)
-int16_t motor_pd_power_main[4];   // モータのスケーリングされた出力格納用(PD制御のための出力)
-int16_t motor_power_main[4];      // 最終的な出力格納用
+int motor_deg[4];             // モータの設置角度を格納用
+int motor_move_power_main[4]; // モータのスケーリングされた出力格納用(移動のための出力)
+int motor_pd_power_main[4];   // モータのスケーリングされた出力格納用(PD制御のための出力)
+int motor_power_main[4];      // 最終的な出力格納用
 
-void motors_init(int16_t deg_1ch, int16_t deg_2ch, int16_t deg_3ch, int16_t deg_4ch)
+void motors_init(int deg_1ch, int deg_2ch, int deg_3ch, int deg_4ch)
 {
     motor_deg[0] = deg_1ch;
     motor_deg[1] = deg_2ch;
@@ -18,7 +18,7 @@ void motors_init(int16_t deg_1ch, int16_t deg_2ch, int16_t deg_3ch, int16_t deg_
     motor_deg[3] = deg_4ch;
 }
 
-void motors_move(int16_t deg, int16_t abs_power)
+void motors_move(int deg, int abs_power)
 {
     abs_power = abs(abs_power); // 一応絶対値にする
 
@@ -26,7 +26,7 @@ void motors_move(int16_t deg, int16_t abs_power)
 
     float sum_x_power = 0; // x方向の出力の合計
     float sum_y_power = 0; // y方向の出力の合計
-    for (int8_t i = 0; i < 4; i++)
+    for (int i = 0; i < 4; i++)
     {
         sum_x_power += get_motor_x_power(i);
         sum_y_power += get_motor_y_power(i);
@@ -37,21 +37,21 @@ void motors_move(int16_t deg, int16_t abs_power)
     if (power_xy > 1e-6)
     {
         float scale_value = float(abs_power) / power_xy;
-        for (int8_t i = 0; i < 4; i++)
+        for (int i = 0; i < 4; i++)
         {
-            motor_move_power_main[i] = int16_t(get_motor_power(i) * scale_value);
+            motor_move_power_main[i] = int(get_motor_power(i) * scale_value);
             motor_move_power_main[i] = constrain(motor_move_power_main[i], -abs_power, abs_power);
         }
     }
     else
     {
-        for (int8_t i = 0; i < 4; i++)
+        for (int i = 0; i < 4; i++)
         {
             motor_move_power_main[i] = 0;
         }
     }
 
-    for (int8_t i = 0; i < 4; i++)
+    for (int i = 0; i < 4; i++)
     {
         motor_power_main[i] = motor_move_power_main[i] + get_PD_power() * motor_pid_sign[i];
         motor_power_main[i] = constrain(motor_power_main[i], -abs_power, abs_power);
@@ -66,30 +66,30 @@ float motor_move_power_sub[4]; // 計算したモータ出力
 float motor_x_power[4];        // モータ出力をx方向に分解
 float motor_y_power[4];        // モータ出力をy方向に分解
 
-void compute_motor_power(int16_t deg, int16_t power)
+void compute_motor_power(int deg, int power)
 {
-    for (int8_t i = 0; i < 4; i++)
+    for (int i = 0; i < 4; i++)
     {
         motor_move_power_sub[i] = sin(radians(deg - motor_deg[i])) * power;
 
-        int16_t power_deg = motor_deg[i] + 90;
+        int power_deg = motor_deg[i] + 90;
 
         motor_x_power[i] = cos(radians(power_deg)) * motor_move_power_sub[i];
         motor_y_power[i] = sin(radians(power_deg)) * motor_move_power_sub[i];
     }
 }
 
-float get_motor_power(int8_t index)
+float get_motor_power(int index)
 {
     return motor_move_power_sub[index];
 }
 
-float get_motor_x_power(int8_t index)
+float get_motor_x_power(int index)
 {
     return motor_x_power[index];
 }
 
-float get_motor_y_power(int8_t index)
+float get_motor_y_power(int index)
 {
     return motor_y_power[index];
 }
